backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
openai==1.51.0
python-docx==1.1.2
PyPDF2==3.0.1
redis==5.1.0
httpx==0.27.2
aiofiles==24.1.0
python-dotenv==1.0.1
aiosqlite==0.20.0
greenlet
